Refactor `if (is_visible())` check with DOD
Inspired by: https://www.youtube.com/watch?v=yy8jQgmhbAU

Performing `if (is_visible())` each time a widget is rendered may result in a heavy amount of cache misses and slow down performance.

Data-oriented design advocates splitting the container from one "items" container into one "visible-items" container and another "non-visible-items" container.

The point of this is to increase performance by reducing cache misses. We know that all the visible items will be stored in the "visible-items" container, so we don't need to check `if (is_visible())` everytime we call render() on a widget.

To do this, we introduce an intermediate ID into items. Showing/hiding items now involves passing this information to the parent AND having the parent move the item into the appropriate container (visible/invisible).

All of this also removes the need to keep a `m_visible` flag to the child and the need for an `is_visible()` function.