Refactor `is_valid()` into "always valid"
Instead of performing `is_valid()` checks during rendering, we ensure that the item is ALWAYS valid.

Fonts and models are two examples that could render an item possibly-invalid. We could have a nullptr font, which makes a text invalid... or a nullptr model, which makes a view invalid. We don't want this.

Currently, two strategies are employed to do so:
 - Add as an argument to the constructor. This way, the user has no choice but to pass a [valid] item. This is used with models. An assert is used to ensure the model is valid.
 - Assign a valid *default* value. This is currently used for fonts. A default font is created in Application's constructor and passed to TextInterface.

When the user wishes to update a possibly invalid parameter, the new value is checked for validity.

All this should [hopefully] ensure that items are always valid and always renderable... and this [hopefully] improves performance.