Solve scenes demo title text issue
Figured out why the title-text didn't show. The title-text ITEM was drawn first before the scene (which is a CANVAS). The ugly solution for now is to iterate through the canvases before the items. Hopefully this is good enough for most cases that we don't need to introduce a z-index to items.